delete profile route and delete avatar when user is deleted
