+++ backend/src/2fa/2fa.controller.ts
@@ -38,7 +38,6 @@ export class TwoFaController {
     const res = await this.twoFaService.verifyTwoFaRoute(id, Number(code));
     await this.twoFaService.deleteTwoFaId(id);
     if (!res || !res.ret) {
-      console.log('res: ', res);
       return { message: 'Invalid two factor code' };
     }
     const tokens = await this.authService.createAndUpdateTokens(res.userId);
