+++ backend/src/app.module.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common';
 import { AuthModule } from './auth/auth.module';
 import { PrismaModule } from './prisma/prisma.module';
 import { UsersModule } from './users/users.module';
-import { ProfileModule } from './profile/profile.module';
 import { ChatGateway } from './chat/chat.gateway';
 import { ChatService } from './chat/chat.service';
 import { StatusModule } from './status/status.module';
@@ -19,7 +18,6 @@ import { ServeStaticModule } from '@nestjs/serve-static';
     UsersModule,
     PrismaModule,
     UsersModule,
-    ProfileModule,
     TwoFaModule,
     StatusModule,
     ServeStaticModule.forRoot({
