--- backend/src/profile/dto/index.ts
@@ -1,1 +1,0 @@
-export * from './profile.dto';
