--- backend/src/profile/profile.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ProfileController } from './profile.controller';
-import { ProfileService } from './profile.service';
-
-describe('ProfileController', () => {
-  //  let controller: ProfileController;
-  //
-  //  beforeEach(async () => {
-  //    const module: TestingModule = await Test.createTestingModule({
-  //      controllers: [ProfileController],
-  //      providers: [ProfileService],
-  //    }).compile();
-  //
-  //    controller = module.get<ProfileController>(ProfileController);
-  //  });
-
-  it('should be defined', () => {
-    expect(true).toBe(true);
-  });
-});
