+++ backend/src/users/users.module.ts
@@ -2,11 +2,12 @@ import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { JwtModule } from '@nestjs/jwt';
+import { TwoFaService } from '../2fa/2fa.service';
 
 @Module({
   imports: [JwtModule.register({})],
   controllers: [UsersController],
-  providers: [UsersService],
+  providers: [UsersService, TwoFaService],
   exports: [UsersService]
 })
 export class UsersModule {}
