+++ backend/src/users/users.service.ts
@@ -1,15 +1,33 @@
 import {
   Injectable,
   NotFoundException,
-  ForbiddenException
+  ForbiddenException,
+  UploadedFile
 } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { JwtService } from '@nestjs/jwt';
 import { UnauthorizedException } from '@nestjs/common';
+import { EditProfileDto } from './dto/profile-user.dto';
+import { TwoFaService } from '../2fa/2fa.service';
+import * as fs from 'fs';
+import { extname } from 'path';
 
 @Injectable()
 export class UsersService {
-  constructor(private prisma: PrismaService, private jwt: JwtService) {}
+  constructor(
+    private prisma: PrismaService,
+    private jwt: JwtService,
+    private twoFa: TwoFaService
+  ) {}
+
+  checkIfUserIsMe(nickname1: string, nickname2: string) {
+    if (nickname1 !== nickname2) {
+      throw new UnauthorizedException(
+        'You are not authorized to perform this action'
+      );
+    }
+  }
+
   async getUser(nickname: string) {
     const user = await this.prisma.user.findUnique({
       where: {
@@ -23,7 +41,6 @@ export class UsersService {
         createdAt: true
       }
     });
-    if (!user) throw new NotFoundException('User not found');
     return user;
   }
 
@@ -40,7 +57,6 @@ export class UsersService {
         createdAt: true
       }
     });
-    if (!user) throw new NotFoundException('User not found');
     return user;
   }
 
@@ -57,63 +73,44 @@ export class UsersService {
     return users;
   }
 
-  //TODO: Use guard to verify authorization
-  async deleteUser(paramName: string, nickname: string) {
-    if (paramName !== nickname) {
-      throw new UnauthorizedException(
-        'You are not authorized to delete this profile'
-      );
-    }
-    //    this.profile.deleteAvatar(avatarPath);
+  async deleteUser(user: object) {
+    this.deleteAvatar(user['avatarPath']);
     const deleteUser = await this.prisma.user.delete({
       where: {
-        nickname: nickname
+        nickname: user['nickname']
       }
     });
     return deleteUser;
   }
 
-  async addFriend(
-    userNickname: string,
-    friendNickname: string,
-    userId: string
-  ) {
-    const user = await this.getUser(userNickname);
-    const friend = await this.getUser(friendNickname);
-    if (userNickname === friendNickname || userId !== user.id) {
+  async addFriend(user: object, friendNickname: string) {
+    if (user['nickname'] === friendNickname) {
       throw new UnauthorizedException(
         'You are not authorized to add this friend'
       );
     }
-    const userFriends = await this.prisma.friend.findUnique({
+    const friend = await this.getUser(friendNickname);
+    if (!friend || !user) throw new NotFoundException('User not found');
+    await this.prisma.friend.upsert({
       where: {
         friend_pkey: {
-          userId: user.id,
+          userId: user['id'],
           friendId: friend.id
         }
       },
-      select: {
-        friendId: true
-      }
-    });
-    if (userFriends) throw new ForbiddenException('Friend already added');
-    await this.prisma.friend.create({
-      data: {
-        userId: user.id,
+      update: {},
+      create: {
+        userId: user['id'],
         friendId: friend.id
       }
     });
     return { message: 'Friend added' };
   }
 
-  async deleteFriend(
-    userNickname: string,
-    friendNickname: string,
-    userId: string
-  ) {
-    const user = await this.getUser(userNickname);
+  async deleteFriend(user: object, friendNickname: string) {
     const friend = await this.getUser(friendNickname);
-    if (userNickname === friendNickname || userId !== user.id) {
+    if (!friend) throw new NotFoundException('Friend not found');
+    if (user['nickname'] === friendNickname) {
       throw new UnauthorizedException(
         'You are not authorized to delete this friend'
       );
@@ -122,7 +119,7 @@ export class UsersService {
       .delete({
         where: {
           friend_pkey: {
-            userId: user.id,
+            userId: user['id'],
             friendId: friend.id
           }
         }
@@ -133,16 +130,10 @@ export class UsersService {
     return { message: 'Friend deleted' };
   }
 
-  async getFriends(userNickname: string, userId: string) {
-    const user = await this.getUser(userNickname);
-    if (userId !== user.id) {
-      throw new ForbiddenException(
-        'You are not authorized to get this user friends'
-      );
-    }
+  async getFriends(user: object) {
     const friendsId = await this.prisma.friend.findMany({
       where: {
-        userId: user.id
+        userId: user['id']
       },
       select: {
         friendId: true
@@ -163,5 +154,85 @@ export class UsersService {
       }
     });
     return friends;
+  }
+
+  async editProfile(userId: string, dto: EditProfileDto) {
+    const user = await this.prisma.user.findUnique({
+      where: {
+        id: userId
+      },
+      select: {
+        twoFactorEnable: true,
+        twoFactorSecret: true
+      }
+    });
+    if (dto.twoFactorEnable && !user.twoFactorEnable) {
+      if (!dto.twoFactorCode) {
+        throw new ForbiddenException('Two factor code required');
+      }
+      const valid = await this.twoFa.verifyTwoFa(
+        user.twoFactorSecret,
+        Number(dto.twoFactorCode)
+      );
+      if (!valid) throw new ForbiddenException('Invalid two factor code');
+    }
+    await this.prisma.user
+      .update({
+        where: {
+          id: userId
+        },
+        data: {
+          nickname: dto.nickname,
+          twoFactorEnable: dto.twoFactorEnable
+        }
+      })
+      .catch(() => {
+        throw new ForbiddenException('Nickname already exists');
+      });
+    return { message: 'Profile updated' };
+  }
+
+  deleteAvatar(avatarPath: string) {
+    if (avatarPath === '/default.jpeg') return;
+    try {
+      fs.unlinkSync(`./public/avatars${avatarPath}`);
+    } catch (err) {
+      console.log(
+        "Can't delete avatar because it doesn't exist (no problem for the user)"
+      );
+    }
+  }
+
+  async uploadAvatar(
+    userId: string,
+    @UploadedFile() file: Express.Multer.File
+  ) {
+    if (!file) throw new ForbiddenException('File required');
+    const allowedTypes = ['.png', '.jpg', '.jpeg'];
+    const fileSizeMb = file.size / 1024 ** 2;
+    const extension = extname(file.originalname);
+    const oldAvatar = await this.prisma.user.findUnique({
+      where: {
+        id: userId
+      },
+      select: {
+        avatarPath: true
+      }
+    });
+    if (!allowedTypes.includes(extension))
+      throw new ForbiddenException('Invalid file type');
+    if (fileSizeMb > 2)
+      throw new ForbiddenException('File size limit exceeded (2MB)');
+    file.path = file.path.replace('public/avatars', '');
+    await this.prisma.user.update({
+      where: {
+        id: userId
+      },
+      data: {
+        avatarPath: file.path
+      }
+    });
+    this.deleteAvatar(oldAvatar.avatarPath);
+    return { message: 'Avatar updated' };
   }
 }
