+++ frontend/src/views/Friends.vue
@@ -89,7 +89,7 @@ export default {
       } catch (error) {
         swall({
           title: 'Error',
-          text: formatError(error.response.data.message),
+          text: formatError(error.message),
           icon: 'error',
           button: 'OK'
         });
@@ -112,7 +112,7 @@ export default {
       } catch (error) {
         swall({
           title: 'Error',
-          text: formatError(error.response.data.message),
+          text: formatError(error.message),
           icon: 'error',
           button: 'OK'
         });
@@ -135,7 +135,7 @@ export default {
       } catch (error) {
         swall({
           title: 'Error',
-          text: formatError(error.response.data.message),
+          text: formatError(error.message),
           icon: 'error',
           button: 'OK'
         });
